requirements.txt
 matplotlib
 tables
 numpy
 scipy
 scikit-learn
 pandas
 mne
 tvb-data
 tvb-library
 tvb-framework